Refactor repo layout and build config
- removed unused files
- lowercase file names
- remade msvc solution and project
