--- BoardGame.sln
@@ -1,30 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.14.36203.30
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "BoardGame", "BoardGame\BoardGame.vcxproj", "{19A76AF3-0481-4CCC-B842-42AE45CA36FF}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{19A76AF3-0481-4CCC-B842-42AE45CA36FF}.Debug|x64.ActiveCfg = Debug|x64
-		{19A76AF3-0481-4CCC-B842-42AE45CA36FF}.Debug|x86.ActiveCfg = Debug|Win32
-		{19A76AF3-0481-4CCC-B842-42AE45CA36FF}.Debug|x86.Build.0 = Debug|Win32
-		{19A76AF3-0481-4CCC-B842-42AE45CA36FF}.Release|x64.ActiveCfg = Release|x64
-		{19A76AF3-0481-4CCC-B842-42AE45CA36FF}.Release|x64.Build.0 = Release|x64
-		{19A76AF3-0481-4CCC-B842-42AE45CA36FF}.Release|x86.ActiveCfg = Release|Win32
-		{19A76AF3-0481-4CCC-B842-42AE45CA36FF}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {24FAB3E0-8459-4EB5-8375-4B52435D172E}
-	EndGlobalSection
-EndGlobal
