--- BoardGame/Entity.cpp
@@ -1,2 +1,0 @@
-#include "Entity.h"
-
