+++ src/entity.h
