+++ vc/snake.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.14.36121.58 d17.14
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "snake", "snake.vcxproj", "{DE726747-E9F3-4EC5-AF95-079BE135692A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{DE726747-E9F3-4EC5-AF95-079BE135692A}.Debug|x64.ActiveCfg = Debug|x64
+		{DE726747-E9F3-4EC5-AF95-079BE135692A}.Debug|x64.Build.0 = Debug|x64
+		{DE726747-E9F3-4EC5-AF95-079BE135692A}.Debug|x86.ActiveCfg = Debug|Win32
+		{DE726747-E9F3-4EC5-AF95-079BE135692A}.Debug|x86.Build.0 = Debug|Win32
+		{DE726747-E9F3-4EC5-AF95-079BE135692A}.Release|x64.ActiveCfg = Release|x64
+		{DE726747-E9F3-4EC5-AF95-079BE135692A}.Release|x64.Build.0 = Release|x64
+		{DE726747-E9F3-4EC5-AF95-079BE135692A}.Release|x86.ActiveCfg = Release|Win32
+		{DE726747-E9F3-4EC5-AF95-079BE135692A}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A74185CC-97AC-4082-AB9A-6D0ED3A8EA51}
+	EndGlobalSection
+EndGlobal
